+++ lib/src/repository.dart
@@ -12,8 +12,8 @@ abstract class QueryBuilder<T> {
   T build(Query query);
 }
 
-class IdentifedObject<T> {
-  IdentifedObject(this.id, this.object);
+class IdentifiedObject<T> {
+  IdentifiedObject(this.id, this.object);
   final String id;
   final T object;
 }
@@ -48,21 +48,30 @@ class RepositoryException implements Exception {
 
 abstract class Repository<T> {
   String? get path;
-
   //read operations
   Future<T> get(String id);
   Stream<T> stream(String id);
   Future<List<T>> query({Query query = const AllQuery()});
   Stream<List<T>> streamQuery({Query query = const AllQuery()});
 
-  //single operations
-  Future<T> add(T item);
-  Future<T> addWithId(String id, T item);
+  Future<T> add(IdentifiedObject<T> item);
   Future<T> update(String id, T Function(T current) updater);
   Future<void> delete(String id);
 
   //batch operations
-  Future<Iterable<T>> addAll(Iterable<T> items);
-  Future<Iterable<T>> updateAll(Iterable<IdentifedObject<T>> items);
+  Future<Iterable<T>> addAll(Iterable<IdentifiedObject<T>> items);
+  Future<Iterable<T>> updateAll(Iterable<IdentifiedObject<T>> items);
   Future<void> deleteAll(Iterable<String> ids);
+
+  //auto identify
+  IdentifiedObject<T> autoIdentify(
+    T object, {
+    T Function(T object, String id)? updateObjectWithId,
+  });
+  Future<T> addAutoIdentified(
+    T object, {
+    T Function(T object, String id)? updateObjectWithId,
+  });
+
+  void dispose();
 }