+++ CHANGELOG.md
@@ -9,3 +9,17 @@
 ## 0.7.0
 
 - added `addWithId` to add an item with a specific id.
+
+## 0.8.0
+
+- Added in memory reference implementation, added a dispose method to the interface 
+- Update interface to make ID generation explicitly an external task thats out of scope of this interface.
+
+## 0.8.1
+
+- Update documentation
+
+## 0.9.0
+
+- Update interface to include auto ID generation - but do not implement any specific ID generation
+- Fix typo in inteface name for IdentifiedObject